Only enqueue CSS if the admin bar is showing
When a logged in user with 'Show Toolbar when viewing site' turned off in their profile visits the front-end, the CSS styles were still being enqueued even though there is no admin bar. This checks to see if the admin bar is showing and if not set returns `should_display` as `false`. (May make the `is_user_logged_in()` check redundant as that is part of `is_admin_bar_showing()`.)